+++ FirebaseAppCheck/Apps/FIRAppCheckTestApp/FIRAppCheckTestApp/AppDelegate.swift
@@ -83,7 +83,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
   func requestLimitedUseToken() {
     AppCheck.appCheck().limitedUseToken { result, error in
       if let result = result {
-        print("FAC limited-use token: \(result.token), expiratoin date: \(result.expirationDate)")
+        print("FAC limited-use token: \(result.token), expiration date: \(result.expirationDate)")
       }
 
       if let error = error {