+++ FirebaseAppCheck/CHANGELOG.md
@@ -1,3 +1,6 @@
+#Unreleased
+- [feature] Added limitedUseTokenWithCompletion() for obtaining limited-use tokens for protecting non-Firebase backends.
+
 # 9.5.0
 - [added] DeviceCheck and App Attest providers are supported by watchOS 9.0+. (#10094, #10098)
 - [added] App Attest provider availability updated to support tvOS 15.0+. (#10093)